rename Video to VideoSlide and add ImageSlide for clarity when profiling
