+++ src/components/VideoSlide.tsx
